Add check for Ray\\Aop\\InterceptedInterface definition
The code has been updated to include a condition that checks if 'Ray\\Aop\\InterceptedInterface' is defined. If it's not defined, a warning will be issued and the function will return false. This addition will help in preventing unexpected errors or crashes that could occur when the 'Ray\\Aop\\InterceptedInterface' is not defined.