Update source files and set CMP0115 policy in CMakeLists.txt
The source files set in the CMakeLists.txt have been updated, where the source file list was move from line 10 to 45 and the file `php_rayaop.h` was added to it. Additionally, the CMake policy CMP0115 has been set to NEW, enabling add_executable() or add_library() to run even if source files are not found, generating an error if the build does not locate the file.